Create links to student/prof pages from the welcome page
