+++ src/main/java/app/models/User.java
@@ -1,10 +1,17 @@
 package app.models;
 
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
 import java.util.ArrayList;
 
 @MappedSuperclass
 public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
 
     private String firstName;
     private String lastName;
@@ -45,6 +52,14 @@ public class User {
     ///////////////
     // Get & Set //
     ///////////////
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public String getFirstName() {
         return firstName;
