+++ src/main/resources/templates/welcome.html
@@ -7,7 +7,17 @@
 </head>
 <body>
     <div sec:authorize="isAuthenticated()">
-        <p>You're logged in. Welcome <span th:text="${currentUser.username}"></span> <button onclick="document.forms['logoutForm'].submit()">Logout</button></p>
+        <p>You're logged in. Welcome <span th:text="${user.firstName}+' '+${user.lastName}+'!'"></span></p>
+        <div sec:authorize="hasRole('ROLE_STUDENT')">
+            <a th:href="'/student/'+${user.id}">Go to student page</a>
+        </div>
+        <div sec:authorize="hasRole('ROLE_PROFESSOR')">
+            <a href="/professor">Go to professor page</a>
+        </div>
+        <div sec:authorize="hasRole('ROLE_COORDINATOR')">
+            <a href="/coordinator">Go to coordinator page</a>
+        </div>
+        <button onclick="document.forms['logoutForm'].submit()">Logout</button>
         <form id="logoutForm" method="POST" action="/logout">
         </form>
     </div>
